examples/saas-dynamic-dashboards/agent/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = ""
authors = ["Orca CopilotKit <testapp@copilotkit.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
copilotkit = "0.1.48"
pydantic-ai-slim = {extras = ["ag-ui"], version = ">=0.0.16"}
openai = ">=1.58.1"
python-dotenv = ">=1.0.1"
langchain = ">=0.1.0"
langchain-core = ">=0.1.5"
langchain-community = ">=0.0.1"
langchain-experimental = ">=0.0.11"
langchain-openai = ">=0.0.1"
langgraph = "^0.3.25"
dotenv = "^0.9.9"
uvicorn = "^0.34.0"
fastapi = "0.115.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "agent:main"
